Adding UserId to Authors and Books.

+++ BookShelf/Services/BookService.cs
@@ -2,6 +2,7 @@
 using BookShelf.Infrastructure.Repositories;
 using BookShelf.Models;
 using BookShelf.Models.BookViewModels;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -9,10 +10,10 @@ namespace BookShelf.Services
 {
     public interface IBookService
     {
-        IEnumerable<BookViewModel> Get();
-        int AddBook(IBookCreateModel model);
-        void UpdateLoanStatus(int id, bool onLoan);
-        void Delete(int id);
+        IEnumerable<BookViewModel> Get(string userId);
+        int AddBook(IBookCreateModel model, string userId);
+        void UpdateLoanStatus(int id, bool onLoan, string userId);
+        void Delete(int id, string userId);
     }
 
     public class BookService : IBookService
@@ -26,33 +27,46 @@ namespace BookShelf.Services
             _authorRepo = authorRepo;
         }
 
-        public IEnumerable<BookViewModel> Get()
+        public IEnumerable<BookViewModel> Get(string userId)
         {
             return _bookRepo.Get()
                 .AsQueryable()
+                .Where(x => x.UserId == userId)
                 .Select(x => BookViewModelFactory.CreateBookViewModel(x));
         }
 
-        public int AddBook(IBookCreateModel model)
+        public int AddBook(IBookCreateModel model, string userId)
         {
             var author = _authorRepo.Get(model.AuthorId);
-            var book = BookFactory.CreateBook(model.Title, author);
+
+            if (author.UserId != userId)
+                throw new Exception("Author could not be found.");
+
+            var book = BookFactory.CreateBook(model.Title, author, userId);
             _bookRepo.Add(book);
             _bookRepo.SaveChanges();
 
             return book.Id;
         }
 
-        public void UpdateLoanStatus(int id, bool onLoan)
+        public void UpdateLoanStatus(int id, bool onLoan, string userId)
         {
             var book = _bookRepo.Get(id);
+
+            if (book.UserId != userId)
+                throw new Exception("Book could not be found.");
+
             book.OnLoan = onLoan;
             _bookRepo.SaveChanges();
         }
 
-        public void Delete(int id)
+        public void Delete(int id, string userId)
         {
             var book = _bookRepo.Get(id);
+
+            if (book.UserId != userId)
+                throw new Exception("Book could not be found.");
+
             _bookRepo.Delete(book);
             _bookRepo.SaveChanges();
         }
